[claude] lift resolver into its own module; NamePath.parse classmethod (#111)
