+++ .uncoded/stubs/src/uncoded/cli.pyi
@@ -3,12 +3,13 @@
 import argparse
 import sys
 from pathlib import Path
-from uncoded.body import SymbolNotFound, UnsupportedNamePath, parse_name_path, resolve_body
+from uncoded.body import resolve_body
 from uncoded.config import find_pyproject_toml, read_instruction_files, read_source_roots
 from uncoded.extract import extract_modules, iter_source_files
 from uncoded.instruction_files import sync_instruction_file
 from uncoded.namespace_map import build_map, render_map
 from uncoded.refs import find_refs
+from uncoded.resolver import NamePath, SymbolNotFound, UnsupportedNamePath
 from uncoded.serena_setup import setup
 from uncoded.skill import sync_skill
 from uncoded.stubs import build_stubs
