+++ src/uncoded/refs.py
@@ -10,8 +10,8 @@ from pathlib import Path
 from typing import IO, cast
 from urllib.parse import unquote, urlparse
 
-from uncoded.body import NamePath, resolve_name_position
 from uncoded.config import find_pyproject_toml
+from uncoded.resolver import NamePath, resolve_name_position
 
 TY_VERSION = "0.0.37"
